Fix for being unable to write to full terminal width
IllegalArgumentExceptions were being thrown when `cursorX + length >= widthInCharacters`,
but this restricts the writable width to widthInCharacters - 1. In the trivial case of
a 1x1 panel, writing a one-character string would always throw. Changed the logic to allow
an additional character of width.

This change will also restore the previous cursor position after a `clear()`. Because
clear() uses write(), which updates the cursor position, a newly-initialized panel
would throw an exception if the user attempted to write to what they would assume
is a default of 0, 0.