Added Maven dependency info to avoid requiring local builds
AsciiPanel is not currently available via Maven Central, so developers are required
to do local builds of the code. This adds a barrier to usage and doesn't
work well in CI environments.

https://jitpack.io/ is a package repository which will resolve Maven coordinates as
GitHub projects and perform the build on behalf of the developer, and works
transparently as a Maven repository serving the associated artifacts.